add flame effect

+++ LEDGeometry/FlameEffect.h
@@ -6,11 +6,13 @@
 namespace LEDGeometry {
 class LEDCurve;
 
+// Adapted from Fire2012 by Mark Kriegsman, July 2012, as part of "Five
+// Elements" shown here: http://youtu.be/knWiGsmgycY
 class FlameEffect : public LightEffect {
    public:
-    FlameEffect(int resolution, int min_y, int max_y);
-    FlameEffect(int cooling, int sparking, int resolution, int min_y,
-                int max_y);
+    FlameEffect(int resolution, float min_y, float max_y);
+    FlameEffect(int cooling, int sparking, int resolution, float min_y,
+                float max_y);
     ~FlameEffect();
 
    protected:
@@ -19,11 +21,11 @@ class FlameEffect : public LightEffect {
     void update_heat();
 
    private:
-    int cooling;
-    int sparking;
+    uint8_t cooling;
+    uint8_t sparking;
     int resolution;
-    int min_y;
-    int max_y;
+    float min_y;
+    float max_y;
     uint8_t* heat;
 };
 }  // namespace LEDGeometry
